requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
tqdm>=4.66.0
loguru>=0.7.0
PyYAML>=6.0.1
